+++ javascript-modules/generate/main.js
@@ -25,6 +25,7 @@ async function run() {
     program.option("--skip-live", "Don't build live editing JS or add live editing scripts to HTML");
     program.option("--skip-components", "Don't build the component browser JS or add component browser scripts to HTML");
     program.option("--disable-bindings", "Don't create data binding panels when live editing");
+    program.option("--editable-regions", "Inject a live editing script that relies on editable regions for rendering. Also disables data binding panels.");
     program.parse(process.argv);
     const options = program.opts();